requirements.txt
asgiref==3.7.2
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==5.0.2
Django==4.2.13
django-admin-interface==0.28.5
django-admin-rangefilter==0.12.1
django-allauth==0.61.1
django-appconf==1.0.6
django-autocomplete-light==3.11.0
django-cleanup==8.1.0
django-colorfield==0.11.0
django-computedfields==0.2.5
django-constance==3.1.0
django-cors-headers==4.3.1
django-cron==0.6.0
django-crum==0.7.9
django-environ==0.11.2
django-extensions==3.2.3
django-fast-update==0.2.3
django-filter==23.5
django-imagekit==5.0.0
django-import-export==3.3.7
django-lifecycle==1.1.2
django-messages-extends==0.6.3
django-model-utils==4.4.0
django-nested-admin==4.0.2
django-picklefield==3.1
django-postgres-extensions==0.9.3
django-sequences==3.0
django-simple-stats==0.7.0
django-smart-selects @ git+https://github.com/iamhayder/django-smart-selects.git@2ff275143389ef649ee57baa063a95e22ea09ff4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
drf-yasg==1.21.7
dynamic-rest==2.3.0
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==21.2.0
hashids==1.3.1
idna==3.6
inflection==0.5.1
MarkupPy==1.14
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pathspec==0.12.1
pilkit==3.0
pillow==10.2.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-monkey-business==1.0.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
xlrd==2.0.1
xlwt==1.3.0
